Assign all g_signal_connect() and friends to gulong variables
Fixes https://bugzilla.gnome.org/show_bug.cgi?id=87821

+++ gtk/gtkcellrenderercombo.c
@@ -40,7 +40,7 @@ struct _GtkCellRendererComboPrivate
 
   gint text_column;
 
-  guint focus_out_id;
+  gulong focus_out_id;
 };
